combine projects and project reducers
