+++ web/src/client/js/reducers/index.js
@@ -2,12 +2,10 @@ import { createStore, applyMiddleware, combineReducers } from 'redux'
 import { composeWithDevTools } from 'redux-devtools-extension/developmentOnly'
 import thunk from 'redux-thunk'
 
-import { project } from './project'
 import { projects } from './projects'
 import { users } from './users'
 
 const rootReducer = combineReducers({
-  project,
   projects,
   users
 })
